% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{higlasso.df}
\alias{higlasso.df}
\title{Synthetic example data for higlasso}
\format{A synthetic data.frame with 500 observations on 11 variables:
\describe{
\item{y}{Continuous response.}
\item{X1-X10}{Covariates.}
}}
\usage{
higlasso.df
}
\description{
X1 and X2 have (degree 3) main effects, and interact with each other.
}
\keyword{datasets}
