% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higlasso.R
\name{higlasso}
\alias{higlasso}
\title{Hierarchical Integrative Group LASSO}
\usage{
higlasso(Y, X, Z, lambda1, lambda2, sigma, degree = 2, maxit = 1000,
  eps = 1e-06, QR = TRUE)
}
\arguments{
\item{Y}{A length n numeric response vector}

\item{X}{A n x p numeric matrix}

\item{Z}{A n x m numeric matrix}

\item{lambda1}{Penalty for main effects}

\item{lambda2}{Penalty for interaction effects}

\item{sigma}{Scale parameter for Integrative weights}

\item{degree}{Degree of \code{bs} basis expansion. Default is 2}

\item{maxit}{Maximum number of iterations. Default is 1000}

\item{eps}{Numeric tolerance for convergence. Defaults to 1e-6}

\item{QR}{Whether or not to QR decompose the design matrix. Default is TRUE}
}
\description{
Hierarchical Integrative Group LASSO
}
\examples{
TODO
}
\author{
Alexander Rix
}
