% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.higlasso.fit}
\alias{predict.higlasso.fit}
\title{Predict via a higlasso fit}
\usage{
\method{predict}{higlasso.fit}(object, newdata, ...)
}
\arguments{
\item{object}{An object of type 'higlasso'}

\item{newdata}{An optional list of length 3, conataining new Y, new Xm, and
new Z. For experts only.}

\item{...}{Additonal parameters to pass on}
}
\description{
Predict via a higlasso fit
}
