% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.higlasso.fit}
\alias{print.higlasso.fit}
\title{Print higlasso fits}
\usage{
\method{print}{higlasso.fit}(x, ...)
}
\arguments{
\item{x}{An object of type 'higlasso'}

\item{...}{Additional parameters to pass onto print}
}
\description{
Print higlasso fits
}
