% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higlasso.R
\name{print.higlasso}
\alias{print.higlasso}
\title{Print higlasso fits}
\usage{
\method{print}{higlasso}(x, ...)
}
\arguments{
\item{x}{An object of type 'higlasso'}

\item{...}{Additional parameters to pass onto print}
}
\description{
Print higlasso fits
}
