% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higlasso.R
\name{summary.higlasso}
\alias{summary.higlasso}
\title{Summarise higlasso fits}
\usage{
\method{summary}{higlasso}(object, ...)
}
\arguments{
\item{object}{An object of type 'higlasso'}

\item{...}{Additional parameters to pass on.}
}
\description{
Summarise higlasso fits
}
