% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{higlasso.df}
\alias{higlasso.df}
\title{Synthetic Example Data For Higlasso}
\format{
A data.frame with 300 observations on 8 variables:
\describe{
\item{Y}{Continuous response.}
\item{V1-V7}{Covariates.}
}
}
\usage{
higlasso.df
}
\description{
This synthetic data is adapted from the linear interaction simulations from
the higlasso paper. The data generating model is:
\deqn{Y = V_1 + V_2 + V_3 + V_7 + V_1 V_2 + V_2 V_7}
}
\keyword{datasets}
