% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higlasso.R
\name{higlasso}
\alias{higlasso}
\title{Hierarchical Integrative Group LASSO}
\usage{
higlasso(Y, X, Z, lambda1, lambda2, sigma = 1, Xm = NULL, degree = 2,
  maxit = 1000, eps = 1e-06, tol = .Machine$double.eps,
  faster = FALSE)
}
\arguments{
\item{Y}{A length n numeric response vector}

\item{X}{A n x p numeric matrix}

\item{Z}{A n x m numeric matrix}

\item{lambda1}{Penalty for main effects}

\item{lambda2}{Penalty for interaction effects}

\item{sigma}{Scale parameter for Integrative weights. Technically a third
tuning paramter but defaults to 1 for computational tractibility}

\item{Xm}{An optional list of design matrices if the user wishes to use their
own groupings instead of a basis expansion}

\item{degree}{Degree of \code{bs} basis expansion. Default is 2}

\item{maxit}{Maximum number of iterations. Default is 1000}

\item{eps}{Numeric tolerance for convergence. Defaults to 1e-6}

\item{tol}{Tolerance for RcppArmadillo solve. This is separate from \code{eps}.
Default is machine epsilon}

\item{faster}{If TRUE, higlasso will be faster, but possibly less accurate.
Default is FALSE}
}
\description{
TODO
}
\details{
We have designed \code{higlasso} to
\itemize{
\item Impose strong heredity constraints on two-way interaction effects
(hierarchical).
\item Incorporate adaptive weights without necessitating initial
coefficient estimates.
\item Induce sparsity for variable selection while respecting group
structure (group LASSO).
}
}
\examples{
TODO
}
\author{
Alexander Rix
}
